requirements.txt
streamlit
newspaper3k
beautifulsoup4
requests
groq
nltk
tqdm